modify (model)Card (router)card.js after card DB redefine

+++ model/Card.js
@@ -9,7 +9,7 @@ const table = `card`;
 const THIS_LOG = `카드`;
 
 const card = {
-    create: (userIdx, cardCompany, cardNumber, cvc, password) => {
+    create: (userIdx, cardCompany, cardNickname, cardNumber, cvc, password) => {
         return new Promise(async (resolve, reject) => {
             const cardEncryptionResult = await encryptionModule.encryption(cardNumber);
             const cvcEncryptionResult = await encryptionModule.encryption(cvc);
@@ -19,8 +19,8 @@ const card = {
             console.log(cvcEncryptionResult);
             console.log(passwordEncryptionResult);
 
-            const cardCreateQuery = `INSERT INTO card(user_idx, card_company_name, card_number, cvc, password, card_salt, cvc_salt, password_salt) VALUES(?, ?, ?, ?, ?, ?, ?, ?)`;
-            const cardCreateResult = await pool.queryParam_Arr(cardCreateQuery, [userIdx, cardCompany, cardEncryptionResult.hashedPassword, cvcEncryptionResult.hashedPassword, passwordEncryptionResult.hashedPassword, cardEncryptionResult.salt, cvcEncryptionResult.salt, passwordEncryptionResult.salt]);
+            const cardCreateQuery = `INSERT INTO card(user_idx, card_company_name, card_nickname, card_number, cvc, card_password, card_salt, cvc_salt, card_password_salt) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?)`;
+            const cardCreateResult = await pool.queryParam_Arr(cardCreateQuery, [userIdx, cardCompany, cardNickname, cardEncryptionResult.hashedPassword, cvcEncryptionResult.hashedPassword, passwordEncryptionResult.hashedPassword, cardEncryptionResult.salt, cvcEncryptionResult.salt, passwordEncryptionResult.salt]);
 
             if (!cardCreateResult) {
                 resolve({
@@ -58,6 +58,7 @@ const card = {
             const cardNumberDecryptionResult = await decryptionModule.decryption(cardData.card_number, cardData.card_salt);
 
             const result = {
+                "cardNickname" : cardData.card_nickname,
                 "cardCompany" : cardData.card_company_name,
                 "cardNumber" : cardNumberDecryptionResult
             };
@@ -75,6 +76,17 @@ const card = {
 
     delete: (userIdx, cardIdx) => {
         return new Promise(async (resolve, reject) => {
+            const userIdxQuery = `SELECT * FROM user WHERE user_idx = ?`;
+            const userIdxResult = await pool.queryParam_Arr(userIdxQuery, [userIdx]);
+
+            if (userIdxResult[0] == undefined) {
+                resolve({
+                    code : statusCode.BAD_REQUEST,
+                    json : authUtil.successFalse(statusCode.BAD_REQUEST, `존재하지 않는 유저입니다`)
+                });
+                return;
+            }
+
             const deleteCardQuery = `DELETE FROM ${table} WHERE user_idx = ? AND card_idx = ?`;
             const deleteCardResult = await pool.queryParam_Arr(deleteCardQuery, [userIdx, cardIdx]);
 
@@ -82,6 +94,14 @@ const card = {
                 resolve({
                     code : statusCode.INTERNAL_SERVER_ERROR,
                     json : authUtil.successFalse(statusCode.INTERNAL_SERVER_ERROR, responseMessage.INTERNAL_SERVER_ERROR)
+                });
+                return;
+            }
+
+            if (deleteCardResult[0] == undefined) {
+                resolve({
+                    code : statusCode.BAD_REQUEST,
+                    json : authUtil.successFalse(statusCode.BAD_REQUEST, `존재하지 않는 카드입니다`)
                 });
                 return;
             }
